app/requirements.txt
aioredis==2.0.1
alembic==1.11.1
anyio==3.7.1
async-timeout==4.0.2
asyncpg==0.28.0
certifi==2023.7.22
cfgv==3.3.1
click==8.1.6
distlib==0.3.7
exceptiongroup==1.1.2
fastapi==0.100.0
filelock==3.12.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
identify==2.5.26
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.1
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
py==1.11.0
pydantic==1.10.11
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-ordering==0.6
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tenacity==8.2.2
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.23.1
virtualenv==20.24.2
